Refactored plugin write queue (v3 plugin architecture).
This causes plugins to write to a new special plugin_wq that contains
prc_plugin_msg_t a structure containing target, command type, and a
message buffer.

prc_msg2() should now be utilized by all plugins, this is a
convenience function that accepts the target, command type, and a
vsnprintf() format-string + varargs, and constructs a prc_plugin_msg_t
appropriately.

The objective is to allow the privmsg handler to modify the message
prior to pushing to the real protocol write queue.

prc_msg() use in plugins is deprecated.