Created new doubly-linked list implementation.
The sll.c was replaced with dll.c. This implements forward/reverse
traversal and enqueue/push/pop operations, in addition to a number of
sanity-checking internal assertions which can be enabled via
-DDEBUG_ASSERT.

dll_pop, unlike sll_pop, now checks if the list is empty, and returns
NULL if this is the case (previously consumers needed to check
themselves).